refactor some functions, handle heredoc in new way
