--- 4.txt
@@ -1,1 +1,0 @@
-      8       8      45
