+++ free.c
@@ -35,12 +35,9 @@ void	free_array(char **array)
 		return ;
 	while (array[i])
 	{
-		if (array[i] != NULL)
-		{
 			free(array[i]);
 			array[i] = NULL;
 			i++;
-		}
 	}
 	free (array);
 	array = NULL;
@@ -91,15 +88,8 @@ void	f_free_cmds(t_cmd *cmds, int cmds_count)
 	cmds = NULL;
 }
 
-void	free_all(t_data *shell)
+void	free_t_data(t_data *shell)
 {
-	if (shell == NULL)
-		return ;
-	if (shell->cmds)
-	{
-		f_free_cmds(shell->cmds, shell->cmds_count);
-		shell->cmds = NULL;
-	}
 	if (shell->pipe_tok)
 	{
 		free_array(shell->pipe_tok);
@@ -110,21 +100,6 @@ void	free_all(t_data *shell)
 		free(shell->input_copy);
 		shell->input_copy = NULL;
 	}
-	if (shell->hd_delimeter)
-	{
-		free (shell->hd_delimeter);
-		shell->hd_delimeter = NULL;
-	}
-	if (shell->tempfile_hd)
-	{
-		free (shell->tempfile_hd);
-		shell->tempfile_hd = NULL;
-	}
-}
-
-void	exit_free(t_data *shell, int status)
-{
-	free_all(shell);
 	if (shell->envp)
 	{
 		free_array(shell->envp);
@@ -135,6 +110,35 @@ void	exit_free(t_data *shell, int status)
 		free_array(shell->new_envp);
 		shell->new_envp = NULL;
 	}
-	get_signal(DEFAULT);
+}
+
+void	free_all(t_data *shell)
+{
+	if (shell == NULL)
+		return ;
+	if (shell->cmds)
+	{
+		f_free_cmds(shell->cmds, shell->cmds_count);
+		shell->cmds = NULL;
+	}
+	free_t_data(shell);
+	if (shell->hd_delimeter)
+	{
+		free (shell->hd_delimeter);
+		shell->hd_delimeter = NULL;
+	}
+	if (shell->tempfile_hd)
+	{
+		free (shell->tempfile_hd);
+		shell->tempfile_hd = NULL;
+	}
+	
+}
+
+void	exit_free(t_data *shell, int status)
+{
+	free_all(shell);
+	
+	get_signal(shell, DEFAULT);
 	exit (status);
 }
