+++ heredoc.c
@@ -32,16 +32,20 @@ void	process_hd(t_data *tokens, const char *file, char *delimeter)
 	int	fd;
 	char	*line;
 
-	get_signal(HEREDOC);
+	get_signal(tokens, HEREDOC);
 	unlink(file);
 	fd = open(file, O_RDWR | O_CREAT | O_TRUNC, 0644);
 	if (fd == -1)
 		error_message(tokens, "Failed to open fd for heredoc", 1);
+	tokens->hd_interrupt = 0;
 	while (1)
 	{
 		line = readline("> ");
-        if (!line)
-            break;
+        if (tokens->hd_interrupt || !line)
+		{
+            free(line);
+			break;
+		}
 		if (ft_strcmp(line, delimeter) == 0)
 		{
 			free(line);
@@ -54,7 +58,7 @@ void	process_hd(t_data *tokens, const char *file, char *delimeter)
 	}
 	if (close(fd) == -1)
 		error_message(tokens, "Failed to close fd for heredoc", 1);
-	get_signal(DEFAULT);
+	get_signal(tokens, DEFAULT);
 	exit (0);
 }
 
@@ -63,9 +67,10 @@ void    *is_heredoc(char *line, t_data *tokens)
 	int		i;
 	int		j;
 	int		len;
-	int		in_quote = 0;
+	int		in_quote;
 	
 	i = 0;
+	in_quote = 0;
 	tokens->hd_count = 0;
 	while (line[i])
 	{
@@ -101,9 +106,30 @@ void    *is_heredoc(char *line, t_data *tokens)
 			printf("HD tempfile %s\n", tokens->tempfile_hd);
 			if (!tokens->tempfile_hd)
 				error_message(tokens, "Failed to assign filename for heredoc", 1);
-			process_hd(tokens, tokens->tempfile_hd, tokens->hd_delimeter);
-            if ( tokens->hd_delimeter != NULL)
-			    free(tokens->hd_delimeter);
+			// process_hd(tokens, tokens->tempfile_hd, tokens->hd_delimeter);
+            // if ( tokens->hd_delimeter != NULL)
+			//     free(tokens->hd_delimeter);
+			pid_t pid = fork();
+            if (pid == 0)
+            {
+                process_hd(tokens, tokens->tempfile_hd, tokens->hd_delimeter);
+                free(tokens->hd_delimeter);
+                exit(0); // Child process exits after processing
+            }
+            else if (pid < 0)
+            {
+                error_message(tokens, "Failed to fork for heredoc processing", 1);
+            }
+            else
+            {
+                // Parent process should wait for the child process to complete
+                int status;
+                waitpid(pid, &status, 0);
+                free(tokens->hd_delimeter);
+                tokens->hd_delimeter = NULL;
+				if (tokens->hd_interrupt == 1)
+					break;
+            }
 		}
         if (line[i] != '\0')
 		    i++;
