+++ structs.h
@@ -17,7 +17,7 @@ typedef enum {
     DEFAULT,
     HANDLER,
     HEREDOC,
-    //NO_SIGNALS
+    NO_SIGNALS
 } t_signal;
 
 typedef enum {
@@ -73,8 +73,9 @@ typedef struct s_data
 	int		redir_count;
 	int		pipe_count;
 	int		**pipe;
-	int *pid;
-	int	exit_status;
+	int		*pid;
+	int		exit_status;
+	int		hd_interrupt;
 	
 	// t_build * shell;
 }	t_data;
